+++ tests/flow/test_fulltext_query.py
@@ -15,7 +15,10 @@ class testFulltextIndexQuery():
         self.graph.query("CALL db.idx.fulltext.createNodeIndex('L3', { field: 'v1', weight: 1 }, { field: 'v2', weight: 2 })")
         self.graph.query("CALL db.idx.fulltext.createNodeIndex('L4', { field: 'v', phonetic: 'dm:en' })")
         self.graph.query("CALL db.idx.fulltext.createNodeIndex('L5', { field: 'v', nostem: true })")
+
+        # create full text index on relationship type E inedxing property 'name'
         self.graph.query("CREATE FULLTEXT INDEX FOR ()-[e:E]-() on (e.name)") 
+
         wait_for_indices_to_sync(self.graph)
 
         n0 = Node(labels="L1", properties={"v": 'hello redis world'})
@@ -24,13 +27,17 @@ class testFulltextIndexQuery():
         n3 = Node(labels="L3", properties={"v1": 'hello redis', "v2": 'hello world'})
         n4 = Node(labels="L4", properties={"v": 'felix'})
         n5 = Node(labels="L5", properties={"v": 'there are seven words in this sentence'})
+
+        # introduce a number of relationships of type E and E1
         e0 = Edge(n5, "E", n0, properties={"name": "just another nice relationship"})
         e1 = Edge(n5, "E", n0, properties={"name": "a nice place to be"})
         e2 = Edge(n5, "E1", n0, properties={"name": "don't find me please, I'm not full text indexed"})
+
+        # create the nodes and relationships
         self.graph.query(f"CREATE {n0}, {n1}, {n2}, {n3}, {n4}, {n5}, {e0}, {e1}, {e2}")
 
-    # full-text query
-    def test01_fulltext_query(self):
+    # test full-text query on nodes
+    def test01_fulltext_node_query(self):
         expected_result = self.graph.query("MATCH (n:L1) RETURN n")
         # fulltext query L1 for hello 
         result = self.graph.query("CALL db.idx.fulltext.queryNodes('L1', 'hello')")
@@ -81,21 +88,45 @@ class testFulltextIndexQuery():
         result = self.graph.query("CALL db.idx.fulltext.queryNodes('L5', 'word')")
         self.env.assertEquals(result.result_set, [])
 
+    # test full-text query on edges
+    def test02_fulltext_edge_query(self):
         # full text query on a relationship E1 (not indexed)
         result = self.graph.query("CALL db.idx.fulltext.queryRelationships('E1', 'please')")
         self.env.assertEquals(result.result_set, [])
 
         # full text query on a relationship E (indexed) 'nice' appears in two relationships
-        result = self.graph.query("CALL db.idx.fulltext.queryRelationships('E', 'nice')")
+        result = self.graph.query("""
+            CALL db.idx.fulltext.queryRelationships('E', 'nice')
+            YIELD relationship AS r
+            RETURN r.name AS name
+            ORDER BY r.name""")
+
+        # expecting two relationships with 'nice' in their name
         self.env.assertEquals(len(result.result_set), 2)
-        actual = [x[0].properties["name"] for x in result.result_set]
-        actual.sort()
+        actual = [row[0] for row in result.result_set]
+
         expected = ["a nice place to be", "just another nice relationship"]
         self.env.assertEquals(actual, expected)
-        # full text query on an indexed label that does not return any match
-        result = self.graph.query("CALL db.idx.fulltext.queryRelationships('E', 'foo')")
+
+        # full text query on an indexed relationship-type that does not return any match
+        result = self.graph.query("CALL db.idx.fulltext.queryRelationships('E', 'nonexistent')")
         self.env.assertEquals(result.result_set, [])
-        # full text query on an indexed label that returns only one of the relationship with that label
-        result = self.graph.query("CALL db.idx.fulltext.queryRelationships('E', 'place')")
+
+        # full-text query on an indexed relationship-type that returns only
+        # a single match
+        result = self.graph.query("""CALL db.idx.fulltext.queryRelationships('E', 'place')
+                                  YIELD relationship AS r
+                                  RETURN r.name""")
         self.env.assertEquals(len(result.result_set), 1)
-        self.env.assertEquals(result.result_set[0][0].properties['name'], "a nice place to be")
+        self.env.assertEquals(result.result_set[0][0], "a nice place to be")
+
+    def test03_fulltext_edge_query_with_crud(self):
+        # this test make sure the index returns valid results
+        # after performing CRUD operations on the indexed entities
+        #
+        # 1. deleting an indexed edge
+        # 2. updateing an indexed edge
+        # 3. adding a new indexed edge
+
+        # TODO: implement this test
+
